chore: move the three all(test, ...)-gated inline test modules to sibling files
